+++ packages/syside-languageserver/scripts/clone-sysml-release.mjs
@@ -28,7 +28,7 @@ const __dirname = path.dirname(__filename);
 const root = path.join(__dirname, "..", "..", "..");
 const dir = path.join(root, "SysML-v2-Release");
 
-const tag = "2024-09";
+const tag = "2024-11";
 
 if (existsSync(dir)) {
     await exec(`git config remote.origin.fetch "+refs/tags/${tag}:refs/tags/${tag}"`, { cwd: dir })